ai_news_bot/pyproject.toml
[project]
name = "ai-news-bot"
version = "0.1.0"
description = "Fetch AI company news and push to Feishu"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "feedparser>=6.0",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "loguru>=0.7",
    "python-dateutil>=2.9",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "respx>=0.21"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["core", "core.fetcher", "core.notifier"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
